API: OrderBy Amount error fixed (SQLite doesn't support this operation with decimal types) Blazor: No Items found text added to TransactionGrid

+++ PersonalFinanceApp.Api/PersonalFinanceApp.Api/Repositories/Implementations/TransactionRepository.cs
@@ -96,7 +96,7 @@ namespace PersonalFinanceApp.Api.Repositories.Implementations
             Expression<Func<Transaction, object>> keySelector = request.SortColumn?.ToLower() switch
             {
                 "location" => t => t.Location!,
-                "amount" => t => t.TotalAmount,
+                "amount" => t => (double)t.TotalAmount,
                 "category" => t => t.CategoryId,
                 "payment method" => t => t.PaymentMethodId,
                 _ => t => t.Date,
